requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
scikit-learn==1.7.1
pandas==2.3.2
numpy==2.3.2
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.11.7
pytest==8.4.1
httpx==0.25.2